Move most methods from 'tractanalysis.tools' to new 'streamline_operations'
